First API design
A simple test exhibiting the draft of an API to use with TARDIS.

What:

Tardis is a non-intrusive cache for applications enabling you to cache
method calls without any modification in your code.

Design:

Creates a proxy to a concrete class on a given namespace. The proxy
fetches or warms up the cache configured for that namespace.
Configuration is tied to a namespace or class name.

Cache implementation is provided by Doctrine.