Creates MethodCall identitfy generator
An interface "IndentityGenerator" is created holding a method
"createIdFor" without any argument signature. Different methods of
identifying something will require different set of arguments.
Although I am not confortable with the current interface, I can't think
of any solution right now.

The `MethodCall` identity generate simply serializes the argument list
of a method call and creates a hash of the method name and the
serialized arguments.
This poses a problem when an argument cannot be serialized, this commit
is me being pragmatic.