Adds methods to identity creation interface
I knew this wouldn't go far. Although some future identity creation for
cache keys may not use only the method name and arguments of a call, a
cache will (as far as I can think of) always belong togueter with a
method call. So having those on the signature of the method makes sense.
Much more sense than leaving it empty.