Proxy factory for public methods of an instance
In order to create a proxy using ProxyManager, we need to list the
proxied class methods we are willing to - wait for it - proxy. This
strategy creates a proxy for all public methods of a class.

Avoids proxying some magic methods as there isn't a way to properly
creating an identity for the proxied values.