An almost decent README
Gives the usage and promise of the library right away: to drink beer.
Holds a (not complete) list of cache storages made availiable by
Doctrine, a FAQ and a list of limitations.

We don't have a fancy logo and neither false promises, which will -
hopefully - be sufficient for people to adopt the library and keep lame
bug reports away from us. Giving us time to drink.