Easier to use Tardis factory API
Since only "MethodCall" identity generator exists, it makes sense to
make it an optional parameter.

Ocramius\ProxyManager is a great component, but we don't need to expose
it to the end user also. So it is optional and we create one instance
that will probably suit 99% of the uses of Tardis. A temporary directory
is used to store proxies and autoload from it.

Static methods are used because there are no class attributes to the
factory. Calling tem statically is better than having to instantiate the
class and calling a method from it.